Read test config in one go and other build workload changes

+++ src/workloads/Openvino_Workload.py
@@ -19,23 +19,13 @@ def is_true(env):
     return value == 'true'
 
 class OpenvinoWorkload(Workload):
-    def update_test_config_from_cmd_line(self, test_config_dict):
-        
-        if pytest.config.getoption('--iterations') != 1:
-            test_config_dict['iterations'] = pytest.config.getoption('iterations')
-
-        if pytest.config.getoption('--exec_mode') != '' and pytest.config.getoption('--exec_mode') != 'None':
-            test_config_dict['exec_mode'] = pytest.config.getoption('exec_mode').split(' ')
-
-        print("\n-- Final Test Configuration Data : \n", test_config_dict)                
-
-
     def install_workload(self, test_config_dict):
         # Installing Openvino within "/home/intel/test/gramine/examples/openvino/openvino_2021"
         install_dir = os.path.join(test_config_dict['workload_home_dir'], 'openvino_2021')
         
         # We would not install if the installation dir already exists.
         if os.path.exists(install_dir):
+            print("\n-- Openvino already installed. Not fetching from source..")
             return
         os.makedirs(install_dir, exist_ok=True)
         cwd = os.getcwd()
@@ -84,11 +74,20 @@ class OpenvinoWorkload(Workload):
     '''
     def replace_manifest_file(self, test_config_dict, buildForThroughput = None):
         original_manifest_file = os.path.join(test_config_dict['workload_home_dir'], test_config_dict['original_manifest_file'])
+        
+        # Check if the original manifest file exists.
+        if not os.path.exists(original_manifest_file):
+            print(f"\n-- Failure: Workload manifest {original_manifest_file} does not exist..")
+            return False
+
+        # Renaming the original manifest to manifest.ori
+        os.rename(original_manifest_file, original_manifest_file+'.ori')
+
         if buildForThroughput == True:
             override_manifest_file = os.path.join(constants.FRAMEWORK_HOME_DIR, test_config_dict['throughput_manifest_file'])
         elif buildForThroughput == False:
             override_manifest_file = os.path.join(constants.FRAMEWORK_HOME_DIR, test_config_dict['latency_manifest_file'])
-        os.rename(original_manifest_file, original_manifest_file+'.ori')
+        
         # Copy the workload specific manifest to workload dir and 
         # rename the same as per the expected original name
         shutil.copy2(override_manifest_file, test_config_dict['workload_home_dir'])
@@ -98,48 +97,66 @@ class OpenvinoWorkload(Workload):
             tmp_original_file = os.path.join(test_config_dict['workload_home_dir'], os.path.basename(test_config_dict['latency_manifest_file']))
         os.rename(tmp_original_file, original_manifest_file)
 
+        return True
+
+
     def build_workload(self, test_config_dict, buildForThroughput = True):
+        if buildForThroughput:
+            if not self.replace_manifest_file(test_config_dict, True):
+                print("\n-- Failure: Workload build failure. Returning without building the workload.")
+                return False
+        else:
+            if not self.replace_manifest_file(test_config_dict, False):
+                print("\n-- Failure: Workload build failure. Returning without building the workload.")
+                return False
+        
         cwd = os.getcwd()
 
-        if buildForThroughput:
-            self.replace_manifest_file(test_config_dict, True)
+        # Not checking for existence of model, as we may be building it for either throughput or latency.
+        # So, this method can be invoked from the caller based on whether we are building the
+        # workload for throughput or latency.
+
+        os.chdir(test_config_dict['workload_home_dir'])
+
+        setupvars_path = os.path.join(test_config_dict['workload_home_dir'], 'openvino_2021/bin', 'setupvars.sh')
+        
+        if os.path.exists(setupvars_path):
+            print(f"\n-- Building workload model '{test_config_dict['model_name']}'..\n")
+            ov_env_var_bld_cmd = f"bash -c 'source {setupvars_path} && make SGX=1'"
+            print("\n-- Setting up OpenVINO environment variables and building Openvino..\n", ov_env_var_bld_cmd)
+            subprocess.run(ov_env_var_bld_cmd, shell=True, check=True)
         else:
-            self.replace_manifest_file(test_config_dict, False)
+            os.chdir(cwd)
+            print(f"\n-- Failure: OpenVino not installed in {setupvars_path}. Returning without building the workload.\n")
+            return False
 
-        if not os.path.exists(test_config_dict['model_dir']):
-            print(f"\n-- Model {test_config_dict['model_name']} is not present. Building the workload model..\n")
+        # After the build is complete we need to del the existing overridden manifest and rename
+        # manifest.ori to the expected original name by the workload.
+        original_manifest_file = os.path.join(test_config_dict['workload_home_dir'], test_config_dict['original_manifest_file'])
+        os.remove(original_manifest_file)
+        os.rename(original_manifest_file+'.ori', original_manifest_file)
 
-            os.chdir(test_config_dict['workload_home_dir'])
-
-            setupvars_path = os.path.join(test_config_dict['workload_home_dir'], 'openvino_2021/bin', 'setupvars.sh')
-            
-            if os.path.exists(setupvars_path):
-                ov_env_var_bld_cmd = f"bash -c 'source {setupvars_path} && make SGX=1'"
-                print("\n-- Setting up OpenVINO environment variables and building Openvino..\n", ov_env_var_bld_cmd)
-                subprocess.run(ov_env_var_bld_cmd, shell=True, check=True)
-            else:
-                os.chdir(cwd)
-                pytest.exit(f"\n-- Failure: OpenVino not installed in {setupvars_path}. Please install the same within the suggested directory and rerun.\n")
-
-            print("\nos.environ['PATH']\n", os.environ['PATH'])
-
-        else:
-            # Model is present. Check for the existence of .xml
-            print(f"\n-- Model {test_config_dict['model_name']} is present..\n")
-
-        #After the build is complete we need to del the existing overridden manifest and rename
-        #manifest.ori to the expected original name by the workload
         os.chdir(cwd)
+
+        # Check for build status by verifying the existence of model.xml file and return accordingly.
+        model_file_path = os.path.join(test_config_dict['model_dir'],test_config_dict['fp'][0],test_config_dict['model_name']+'.xml')
+        if os.path.exists(model_file_path):
+            print("\n-- Model is built. Can proceed further to execute the workload..")
+            return True
+        else:
+            print("\n-- Failure: Model not found/built. Returning without building the model.")
+            return False
 
 
     def pre_actions(self, test_config_dict):
+        utils.set_cpu_freq_scaling_governor()        
         self.install_workload(test_config_dict)
+        
 
+    def execute_workload(self, test_config_dict, execForThroughput = True):
+        pass
 
     def __init__(self, test_config_dict):
-        print("\nprinting self Openvino workload_obj\n")
-        #self.preexisting_pids = psutil.pids()
-
         # result is the overall score for all benchmarks
         # it is a list containing [baseline_score, testapp_score, percent_degradaton]
         self.result = []
